phase-07-cicd-production/requirements.txt
flask==3.0.3
psycopg2-binary==2.9.9
prometheus-client==0.19.0

